% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdbgdr}
\alias{cdbgdr}
\title{Community Development Block Grant - DR}
\format{
An object of class \code{data.frame} with 13843 rows and 727 columns.
}
\source{
\url{https://www.hudexchange.info/programs/cdbg-dr/}
}
\usage{
cdbgdr
}
\description{
"The Community Development Block Grant,
one of the longest-running programs of
the U.S. Department of Housing and Urban Development,
funds local community development activities
with the stated goal of providing affordable housing,
anti-poverty programs, and infrastructure development.(Wikipedia)"

A data frame with 727 columns and 17492 rows.
}
\keyword{datasets}
