% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dd}
\alias{dd}
\title{Disaster Declarations Summaries}
\format{
An object of class \code{data.frame} with 61016 rows and 19 columns.
}
\source{
\url{https://www.fema.gov/openfema-data-page/disaster-declarations-summaries-v1}
}
\usage{
dd
}
\description{
"FEMA Disaster Declarations Summary is a summarized dataset describing all
federally declared disasters. This dataset lists all official FEMA Disaster Declarations,
beginning with the first disaster declaration in 1953 and features all three disaster
declaration types: major disaster, emergency, and fire management assistance. The
dataset includes declared recovery programs and geographic areas
(county not available before 1964; Fire Management records are considered partial
due to historical nature of the dataset)."
}
\keyword{datasets}
