% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hma}
\alias{hma}
\title{Hazard Mitigation Assistance Projects}
\format{
An object of class \code{data.frame} with 29769 rows and 30 columns.
}
\source{
\url{https://www.fema.gov/openfema-data-page/hazard-mitigation-assistance-projects-v2}
}
\usage{
hma
}
\description{
"Hazard mitigation planning reduces loss of life and property by minimizing
the impact of disasters.
It begins with state, tribal and local governments identifying
natural disaster risks and vulnerabilities that are common in their area.
After identifying these risks, they develop long-term strategies for protecting
people and property from similar events. Mitigation plans are key
to breaking the cycle of disaster damage and reconstruction."
}
\keyword{datasets}
