% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ihp}
\alias{ihp}
\title{Individuals and Household Programs Valid Registrations}
\format{
An object of class \code{data.frame} with 9491 rows and 66 columns.
}
\source{
\url{https://www.fema.gov/openfema-data-page/individuals-and-households-program-valid-registrations-v1}
}
\usage{
ihp
}
\description{
"IHP is intended to meet basic needs and supplement disaster recovery
efforts. The IHP is not intended to return disaster-damaged property
to its pre-disaster condition. Disaster damage to secondary or vacation
homes does not qualify for IHP assistance."

Live births data from census.gov.
}
\keyword{datasets}
