% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{noaastorm}
\alias{noaastorm}
\title{National Oceanic and Atmospheric Administration - Storm Events}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1333460 rows and 52 columns.
}
\source{
\url{https://www.ncdc.noaa.gov/stormevents/}
}
\usage{
noaastorm
}
\description{
"The occurrence of storms and other significant weather phenomena having
sufficient intensity to cause loss of life, injuries, significant property
damage, and/or disruption to commerce;
Rare, unusual, weather phenomena that generate media attention,
such as snow flurries in South Florida or the San Diego coastal area; and
Other significant meteorological events, such as record maximum or minimum
temperatures or precipitation that occur in connection with another event."
}
\keyword{datasets}
