% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pa}
\alias{pa}
\title{Public Assistance Funded Projects Detail}
\format{
An object of class \code{data.frame} with 728906 rows and 22 columns.
}
\source{
\url{https://www.fema.gov/openfema-data-page/public-assistance-funded-projects-details-v1}
}
\usage{
pa
}
\description{
"FEMA provides supplemental Federal disaster grant assistance for
debris removal, emergency protective measures, and the repair,
replacement, or restoration of disaster-damaged, publicly owned
facilities and the facilities of certain Private Non-Profit (PNP)
organizations through the PA Program (CDFA Number 97.036).
The PA Program also encourages protection of these damaged
facilities from future events by providing assistance for
hazard mitigation measures during the recovery process."
}
\keyword{datasets}
